Create README.md
I have created a webpage for a basic code editor using HTML, CSS, and JavaScript. In the code editor, I have included a copy button, a save button, and a lock/unlock button. You can use the copy button to copy all the code with one click, and the save button to save it. The purpose of the lock/unlock button is to secure your code editor so that no one can make changes to your code.